tests: Add replacement policy tests
Add tests to test the correctness of replacement policies using TrafficGen.
Enable debug-flags to print the hit and miss messages so that you can
compare the results with the comments in the tests. Even though the tests
are targeting specific replacement policies, they can be reused to test all
replacement policies.

Change-Id: I3a8013fbcb19adae25b0818ac9e4b0be60be0689